Reorganize SQLAlchemy version-dependent code
