+++ sqlakeyset/sqla14.py
@@ -1,17 +1,31 @@
 """Methods for messing with the internals of SQLAlchemy >1.3 results."""
-from sqlalchemy.engine.row import Row as _Row, LegacyRow as _LegacyRow
+from sqlalchemy.engine.row import LegacyRow as _LegacyRow
 
-from .sqla14_aux import *
+from .constants import ORDER_COL_PREFIX
+from .sqla20 import (
+    Row,
+    core_result_type,
+    group_by_clauses,
+    order_by_clauses,
+    orm_coerce_row,
+    orm_query_keys,
+    orm_result_type,
+    orm_to_selectable,
+    result_keys,
+)
 
 
-class LegacyRow(_Row):
+class LegacyRow(_LegacyRow):
     def keys(self):
         return result_keys(self._parent)
 
 
 def core_coerce_row(row, extra_columns, result_type):
     """Trim off the extra columns and return as a correct-as-possible
-    sqlalchemy Row."""
+    sqlalchemy Row.
+
+    In SQLAlchemy 1.4 there is the possibility of the row being a transitional
+    LegacyRow, so we handle this case explicitly."""
     if not extra_columns:
         return row
     N = len(row) - len(extra_columns)
